requirements.txt
django
djangorestframework
paramiko
xhtml2pdf
xlwt
requests
django-session-timeout
numpy
mysqlclient